requirements.txt
asgiref
Django
django-cors-headers
djangorestframework
install
numpy
pandas
python-dateutil
pytz
six
sqlparse
tzdata
